Revert "Idva5 1348 data to be stored on data wrapper"
