+++ src/test/java/uk/gov/companieshouse/acsp/service/AcspServiceTest.java
@@ -14,9 +14,7 @@ import uk.gov.companieshouse.acsp.exception.ServiceException;
 import uk.gov.companieshouse.acsp.exception.SubmissionNotLinkedToTransactionException;
 import uk.gov.companieshouse.acsp.models.dao.AcspDataDao;
 import uk.gov.companieshouse.acsp.models.dao.AcspDataSubmissionDao;
-import uk.gov.companieshouse.acsp.models.dao.DataDao;
 import uk.gov.companieshouse.acsp.models.dto.AcspDataDto;
-import uk.gov.companieshouse.acsp.models.dto.AcspDataWrapperDto;
 import uk.gov.companieshouse.acsp.models.dto.CompanyDto;
 import uk.gov.companieshouse.acsp.repositories.AcspRepository;
 import uk.gov.companieshouse.acsp.util.TransactionUtils;
@@ -92,26 +90,16 @@ class AcspServiceTest {
 
     @Test
     void createAcspSuccess() throws Exception {
+        AcspDataDto acspData = new AcspDataDto();
+        acspData.setId("demo@ch.gov.uk");
 
-        AcspDataDto acspDataDto = new AcspDataDto();
-        acspDataDto.setId("demo@ch.gov.uk");
-
-        DataDao data  = new DataDao();
         AcspDataDao acspDataDao = new AcspDataDao();
-        acspDataDao.setData(data);
         acspDataDao.setId("demo@ch.gov.uk");
-
-        AcspDataWrapperDto acspDataWrapperDto = new AcspDataWrapperDto();
-        acspDataWrapperDto.setId("demo@ch.gov.uk");
-        acspDataWrapperDto.setData(acspDataDto);
-
-
-
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.insert((AcspDataDao) any())).thenReturn(acspDataDao);
         doNothing().when(transactionService).updateTransaction(any(), any());
         ResponseEntity<Object> response = acspService.createAcspRegData(transaction,
-                acspDataDto,
+                acspData,
                 REQUEST_ID);
         assertEquals(HttpStatus.CREATED, response.getStatusCode());
     }
@@ -123,12 +111,7 @@ class AcspServiceTest {
 
         AcspDataDao acspDataDao = new AcspDataDao();
         acspDataDao.setId("demo@ch.gov.uk");
-
-        AcspDataWrapperDto acspDataWrapperDto = new AcspDataWrapperDto();
-        acspDataWrapperDto.setData(acspData);
-        acspDataWrapperDto.setId("demo@ch.gov.uk");
-
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.insert((AcspDataDao) any())).thenThrow(DuplicateKeyException.class);
         ResponseEntity<Object> response = acspService.createAcspRegData(transaction,
                 acspData,
@@ -138,44 +121,34 @@ class AcspServiceTest {
 
     @Test
     void createAcspException() {
-        AcspDataDto acspDataDto = new AcspDataDto();
-        acspDataDto.setId("demo@ch.gov.uk");
+        AcspDataDto acspData = new AcspDataDto();
+        acspData.setId("demo@ch.gov.uk");
 
         AcspDataDao acspDataDao = new AcspDataDao();
         acspDataDao.setId("demo@ch.gov.uk");
-
-        AcspDataWrapperDto acspDataWrapperDto = new AcspDataWrapperDto();
-        acspDataWrapperDto.setData(acspDataDto);
-        acspDataWrapperDto.setId("demo@ch.gov.uk");
-
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.insert((AcspDataDao) any())).thenThrow(MongoSocketWriteException.class);
         ResponseEntity<Object> response = acspService.createAcspRegData(transaction,
-                acspDataDto,
+                acspData,
                 REQUEST_ID);
         assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, response.getStatusCode());
     }
 
     @Test
     void updateAcsp() throws Exception {
-        AcspDataDto acspDataDto = new AcspDataDto();
-        acspDataDto.setId("demo@ch.gov.uk");
+        AcspDataDto acspData = new AcspDataDto();
+        acspData.setId("demo@ch.gov.uk");
 
         AcspDataDao acspDataDao = new AcspDataDao();
         acspDataDao.setId("demo@ch.gov.uk");
-
-        AcspDataWrapperDto acspDataWrapperDto = new AcspDataWrapperDto();
-        acspDataWrapperDto.setData(acspDataDto);
-        acspDataWrapperDto.setId("demo@ch.gov.uk");
-
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
         ResponseEntity<Object> response = acspService.updateACSPDetails(transaction,
-                acspDataDto,
+                acspData,
                 REQUEST_ID,
                 USER_ID);
         assertEquals(HttpStatus.OK, response.getStatusCode());
@@ -189,10 +162,9 @@ class AcspServiceTest {
 
         AcspDataDao acspDataDao = new AcspDataDao();
         acspDataDao.setId("demo@ch.gov.uk");
-
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
         when(transaction.getStatus()).thenReturn(TransactionStatus.CLOSED_PENDING_PAYMENT);
 
@@ -261,16 +233,12 @@ class AcspServiceTest {
 
     @Test
     void getAcspApplicationStatusOpenTransaction() throws ServiceException {
-
         var application = new AcspDataDao();
-        var dataDao = new DataDao();
-        application.setData(dataDao);
         var submissionData = new AcspDataSubmissionDao();
-
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        dataDao.setLinks(links);
-        dataDao.setAcspDataSubmission(submissionData);
+        application.setLinks(links);
+        application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.NO_CONTENT);
         Transaction transaction = new Transaction();
         transaction.setStatus(TransactionStatus.OPEN);
@@ -284,14 +252,11 @@ class AcspServiceTest {
     @Test
     void getAcspApplicationStatusTransactionClosedAndRejected() throws ServiceException {
         var application = new AcspDataDao();
-        var dataDao = new DataDao();
-        application.setData(dataDao);
         var submissionData = new AcspDataSubmissionDao();
-
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        dataDao.setLinks(links);
-        dataDao.setAcspDataSubmission(submissionData);
+        application.setLinks(links);
+        application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.NOT_FOUND);
         Transaction transaction = new Transaction();
         transaction.setStatus(TransactionStatus.CLOSED);
@@ -311,14 +276,11 @@ class AcspServiceTest {
     @Test
     void getAcspApplicationStatusTransactionClosedAndProcessing() throws ServiceException {
         var application = new AcspDataDao();
-        var dataDao = new DataDao();
-        application.setData(dataDao);
         var submissionData = new AcspDataSubmissionDao();
-
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        dataDao.setLinks(links);
-        dataDao.setAcspDataSubmission(submissionData);
+        application.setLinks(links);
+        application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.FORBIDDEN);
         Transaction transaction = new Transaction();
         transaction.setStatus(TransactionStatus.CLOSED);
@@ -337,14 +299,11 @@ class AcspServiceTest {
     @Test
     void getAcspApplicationStatusError() throws ServiceException {
         var application = new AcspDataDao();
-        var dataDao = new DataDao();
-        application.setData(dataDao);
         var submissionData = new AcspDataSubmissionDao();
-
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        dataDao.setLinks(links);
-        dataDao.setAcspDataSubmission(submissionData);
+        application.setLinks(links);
+        application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 
         when(acspRepository.findById(USER_ID)).thenReturn(Optional.of(application));
@@ -359,7 +318,7 @@ class AcspServiceTest {
         acspDataDao.setId(SUBMISSION_ID);
         when(acspRepository.findById(SUBMISSION_ID)
         ).thenReturn(Optional.of(acspDataDao));
-        when(acspRegDataDtoDaoMapper.daoToDto(acspDataDao.getData())
+        when(acspRegDataDtoDaoMapper.daoToDto(acspDataDao)
         ).thenReturn(acspDataDto);
 
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
@@ -373,7 +332,7 @@ class AcspServiceTest {
         var acspDataDao = new AcspDataDao();
         acspDataDao.setId(SUBMISSION_ID);
         when(acspRepository.findById(SUBMISSION_ID)).thenReturn(Optional.of(acspDataDao));
-        when(acspRegDataDtoDaoMapper.daoToDto(acspDataDao.getData())).thenReturn(acspDataDto);
+        when(acspRegDataDtoDaoMapper.daoToDto(acspDataDao)).thenReturn(acspDataDto);
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(false);
         transaction.setId("1234");
@@ -416,7 +375,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -443,7 +402,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -470,7 +429,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -497,7 +456,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -526,7 +485,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -553,7 +512,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -578,7 +537,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
@@ -602,7 +561,7 @@ class AcspServiceTest {
         acspDataDao.setId("demo@ch.gov.uk");
         when(transactionUtils.isTransactionLinkedToAcspSubmission(eq(transaction), any(AcspDataDto.class)))
                 .thenReturn(true);
-        when(acspRegDataDtoDaoMapper.dtoToDao(any())).thenReturn(acspDataDao);
+        when(acspRegDataDtoDaoMapper.dtoToDao(acspData)).thenReturn(acspDataDao);
         when(acspRepository.save(any())).thenReturn(acspDataDao);
 
         when(transaction.getStatus()).thenReturn(TransactionStatus.OPEN);
